# padariaSwagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "PADOCA",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/estoque": {
      "get": {
        "tags": [
          "Estoque"
        ],
        "description": "Endpoint (procedure) para obter todos os produtos no estoques da padaria.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Estoque"
        ],
        "description": "Endpoint para cadastrar um produto no estoque.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Objeto para criação de Estoque",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idProduto": {
                  "type": "number",
                  "example": 1
                },
                "quantidade": {
                  "type": "number",
                  "example": 100
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/estoque/{id}": {
      "get": {
        "tags": [
          "Estoque"
        ],
        "description": "Endpoint trazer apenas um produto por id no Estoque.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "in": "params",
          "required": true
        }
      },
      "put": {
        "tags": [
          "Estoque"
        ],
        "description": "Endpoint para atualizar um produto do Estoque.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Objeto para criação de Estoque",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idProduto": {
                  "type": "number",
                  "example": 1
                },
                "quantidade": {
                  "type": "number",
                  "example": 100
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/produto": {
      "get": {
        "tags": [
          "Produto"
        ],
        "description": "Endpoint (procedure) para obter todos os produtos da padaria.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Produto"
        ],
        "description": "Endpoint para cadastrar um produto.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Objeto para criação de Produto",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Coca-cola Lata 350 ml."
                },
                "marca": {
                  "type": "string",
                  "example": "Coca"
                },
                "precoVenda": {
                  "type": "number",
                  "example": 2.8
                },
                "bitAtivo": {
                  "type": "number",
                  "example": 1
                },
                "quantidade": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/produto/{id}": {
      "get": {
        "tags": [
          "Produto"
        ],
        "description": "Endpoint para trazer um produto pelo id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Produto"
        ],
        "description": "Endpoint para atualizar um Produto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Objeto para criação de Produto",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Coca-cola Lata 350 ml."
                },
                "marca": {
                  "type": "string",
                  "example": "Coca"
                },
                "precoVenda": {
                  "type": "number",
                  "example": 2.8
                },
                "bitAtivo": {
                  "type": "number",
                  "example": 1
                },
                "quantidade": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/venda": {
      "get": {
        "tags": [
          "Venda"
        ],
        "description": "Endpoint para obter todos os vendas da padaria.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Venda"
        ],
        "description": "Endpoint para realizar uma venda.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Objeto para criação de venda",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idProduto": {
                  "type": "number",
                  "example": 1
                },
                "valor": {
                  "type": "number",
                  "example": 3.5
                },
                "quantidade": {
                  "type": "number",
                  "example": 2
                },
                "data": {
                  "type": "string",
                  "example": "2022-12-05"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/venda/{id}": {
      "get": {
        "tags": [
          "Venda"
        ],
        "description": "Endpoint para trazer a venda pelo id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Venda"
        ],
        "description": "Endpoint para atualizar uma venda.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Objeto para criação de venda",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idProduto": {
                  "type": "number",
                  "example": 1
                },
                "valor": {
                  "type": "number",
                  "example": 3.5
                },
                "quantidade": {
                  "type": "number",
                  "example": 2
                },
                "data": {
                  "type": "string",
                  "example": "2022-12-05"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  }
}